runner: improve description when entity is not found
The Rust standard library uses "entity not found" for POSIX errno code
2 for when there is no such file or directory.  This confuses users
and isn't immediately obvious when the error from geckodriver reads
"unknown error: entity not found".

This patch overrides the `ErrorKind::NotFound` description with "no such
file or directory", which is the de facto standard elsewhere until a
suitable fix has been upstream to Rust.

It also returns the wrapped error as the cause and _not_ the cause of
the wrapped error.