build: print c compiler version
The command `cc' may be symlinked through /etc/alternatives on Debian
systems.  If the CC output variable is not set, cc is used and it may
not resolve to gcc if the system default is clang.

Since clang is known to cause problems with Rust toolchains, printing
the cc version at the beginning of the build script will help diagnose
the issue.